+++ web/pgadmin/browser/server_groups/servers/databases/schemas/tables/__init__.py
@@ -593,6 +593,8 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         status, res = self._fetch_properties(did, scid, tid)
         if not status:
             return res
+        if not res['rows']:
+            return gone(gettext("The specified table could not be found."))
 
         return super(TableView, self).properties(
             gid, sid, did, scid, tid, res
@@ -1151,6 +1153,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
             if not status:
                 return internal_server_error(errormsg=res)
 
+            if len(res['rows']) == 0:
+                return gone(gettext("The specified table could not be found."))
+
             return super(TableView, self).truncate(
                 gid, sid, did, scid, tid, res
             )
@@ -1370,6 +1375,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         if not status:
             return res
 
+        if len(res['rows']) == 0:
+            return gone(gettext("The specified table could not be found."))
+
         data = res['rows'][0]
 
         return BaseTableView.get_reverse_engineered_sql(
@@ -1398,6 +1406,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         status, res = self.conn.execute_dict(SQL)
         if not status:
             return internal_server_error(errormsg=res)
+
+        if len(res['rows']) == 0:
+            return gone(gettext("The specified table could not be found."))
 
         data = res['rows'][0]
         data = self._formatter(did, scid, tid, data)
@@ -1443,6 +1454,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         status, res = self.conn.execute_dict(SQL)
         if not status:
             return internal_server_error(errormsg=res)
+
+        if len(res['rows']) == 0:
+            return gone(gettext("The specified table could not be found."))
 
         data = res['rows'][0]
         data = self._formatter(did, scid, tid, data)
@@ -1491,6 +1505,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         status, res = self.conn.execute_dict(SQL)
         if not status:
             return internal_server_error(errormsg=res)
+
+        if len(res['rows']) == 0:
+            return gone(gettext("The specified table could not be found."))
 
         data = res['rows'][0]
         data = self._formatter(did, scid, tid, data)
@@ -1542,6 +1559,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         if not status:
             return internal_server_error(errormsg=res)
 
+        if len(res['rows']) == 0:
+            return gone(gettext("The specified table could not be found."))
+
         data = res['rows'][0]
 
         sql = u"DELETE FROM {0}\n\tWHERE <condition>;".format(
@@ -1587,6 +1607,9 @@ class TableView(BaseTableView, DataTypeReader, VacuumSettings,
         data = {}
         data['schema'], data['name'] = \
             super(TableView, self).get_schema_and_table_name(tid)
+
+        if data['name'] is None:
+            return gone(gettext("The specified table could not be found."))
 
         SQL = render_template(
             "/".join(
